+++ app/src/main/java/ru/s4m1d/notes/app/desktop/client/app/runner/AppStarter.java
@@ -7,10 +7,10 @@ import org.springframework.context.annotation.Configuration;
 import ru.s4m1d.notes.app.desktop.client.model.context.SwingApplicationContext;
 import ru.s4m1d.notes.app.desktop.client.model.context.NotesContext;
 import ru.s4m1d.notes.app.desktop.client.model.context.WorkspaceContext;
-import ru.s4m1d.notes.app.desktop.client.daemon.task.MainFrameStateService;
+import ru.s4m1d.notes.app.desktop.client.presenter.mainframe.MainFrameStateService;
 import ru.s4m1d.notes.app.desktop.client.daemon.task.context.ContextInitializationTask;
-import ru.s4m1d.notes.app.desktop.client.daemon.notes.NotesMonitoringServiceImpl;
-import ru.s4m1d.notes.app.desktop.client.daemon.notes.NotesBarUpdateServiceImpl;
+import ru.s4m1d.notes.app.desktop.client.presenter.notes.NotesMonitoringServiceImpl;
+import ru.s4m1d.notes.app.desktop.client.presenter.notes.NotesBarUpdateServiceImpl;
 import ru.s4m1d.notes.app.desktop.client.daemon.task.workspace.WorkspaceObserver;
 import ru.s4m1d.notes.app.desktop.client.daemon.task.workspace.WorkspaceService;
 import ru.s4m1d.notes.app.desktop.client.daemon.task.workspace.WorkspaceWorker;
