--- daemon-api/src/main/java/ru/s4m1d/notes/app/desktop/client/daemon/task/NotesSearchService.java
@@ -1,9 +1,0 @@
-package ru.s4m1d.notes.app.desktop.client.daemon.task;
-
-import ru.s4m1d.notes.app.desktop.client.model.note.Note;
-
-import java.util.List;
-
-public interface NotesSearchService {
-    public List<Note> searchForNotes();
-}
