+++ daemon-impl/pom.xml
@@ -42,6 +42,11 @@
             <artifactId>util</artifactId>
             <version>${parent.version}</version>
         </dependency>
+        <dependency>
+            <groupId>ru.s4m1d</groupId>
+            <artifactId>presenter</artifactId>
+            <version>${parent.version}</version>
+        </dependency>
     </dependencies>
 
 </project>
