--- daemon-impl/src/main/java/ru/s4m1d/notes/app/desktop/client/daemon/context/service/ContextService.java
@@ -1,15 +1,0 @@
-package ru.s4m1d.notes.app.desktop.client.daemon.context.service;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.Setter;
-import ru.s4m1d.notes.app.desktop.client.model.context.SwingApplicationContext;
-
-@AllArgsConstructor
-@Builder
-@Getter
-@Setter
-public class ContextService {
-    private SwingApplicationContext swingApplicationContext;
-}
