--- daemon-impl/src/main/java/ru/s4m1d/notes/app/desktop/client/daemon/task/NewComponentEvent.java
@@ -1,6 +1,0 @@
-package ru.s4m1d.notes.app.desktop.client.daemon.task;
-
-import ru.s4m1d.notes.app.desktop.client.core.Event;
-
-public class NewComponentEvent implements Event {
-}
