+++ presenter-api/pom.xml
@@ -9,7 +9,7 @@
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>app</artifactId>
+    <artifactId>presenter-api</artifactId>
 
     <properties>
         <maven.compiler.source>11</maven.compiler.source>
@@ -19,17 +19,7 @@
     <dependencies>
         <dependency>
             <groupId>ru.s4m1d</groupId>
-            <artifactId>components</artifactId>
-            <version>${parent.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>ru.s4m1d</groupId>
-            <artifactId>daemon-impl</artifactId>
-            <version>${parent.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>ru.s4m1d</groupId>
-            <artifactId>presenter</artifactId>
+            <artifactId>model</artifactId>
             <version>${parent.version}</version>
         </dependency>
     </dependencies>
