+++ presenter/src/main/java/ru/s4m1d/notes/app/desktop/client/presenter/mainframe/NewComponentEvent.java
@@ -1,0 +1,6 @@
+package ru.s4m1d.notes.app.desktop.client.presenter.mainframe;
+
+import ru.s4m1d.notes.app.desktop.client.core.Event;
+
+public class NewComponentEvent implements Event {
+}
