+++ presenter/src/main/java/ru/s4m1d/notes/app/desktop/client/presenter/notes/mapper/FilesToNotesMapper.java
@@ -1,0 +1,23 @@
+package ru.s4m1d.notes.app.desktop.client.presenter.notes.mapper;
+
+import ru.s4m1d.notes.app.desktop.client.model.content.FileContent;
+import ru.s4m1d.notes.app.desktop.client.model.note.Note;
+import ru.s4m1d.notes.app.desktop.client.util.StringUtil;
+
+import java.io.File;
+import java.util.ArrayList;
+import java.util.List;
+
+public class FilesToNotesMapper {
+    public List<Note> map(List<File> files) {
+        List<Note> result = new ArrayList<>();
+        for (File file: files) {
+            Note note = Note.builder()
+                    .name(StringUtil.cutExtension(file.getName()))
+                    .content(new FileContent(file))
+                    .build();
+            result.add(note);
+        }
+        return result;
+    }
+}
